LCD, LM35DZ, and PH-4502C.

+++ include/module_lm35dz.h
@@ -1,6 +1,6 @@
 #ifndef MODULE_LM35DZ_H
 #define MODULE_LM35DZ_H
 
-int read_lm35dz();
+float read_lm35dz();
 
 #endif
